backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cassandra-perplexity-mcp"
version = "0.1.0"
description = "Perplexity MCP server — web search and grounded AI answers via Sonar API"
requires-python = ">=3.12"
dependencies = [
    "cassandra-mcp-auth @ git+https://github.com/Cassandras-Edge/cassandra-auth.git@8c8d8e7#subdirectory=python",
    "cryptography>=44.0.0",
    "fastmcp>=3.2.0",
    "httpx>=0.28.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
cassandra-perplexity-mcp = "cassandra_perplexity_mcp.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
